+++ packages/terra-slide-panel/CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Fixed
+* Setting Focus to slide panel's panel content on open.
 
 3.2.0 - (March 26, 2019)
 ------------------